Moved all omdb actions to omdb_service, updated tests to support update.
