+++ requirements-test.txt
@@ -1,29 +1,66 @@
 #
-# This file is autogenerated by pip-compile
+# This file is autogenerated by pip-compile with python 3.9
 # To update, run:
 #
 #    pip-compile --output-file=requirements-test.txt requirements-test.in
 #
-atomicwrites==1.3.0       # via pytest
-attrs==19.1.0             # via packaging, pytest
-colorama==0.4.1           # via pytest
-coverage==4.5.4           # via pytest-cov
+atomicwrites==1.3.0
+    # via pytest
+attrs==19.1.0
+    # via
+    #   packaging
+    #   pytest
+colorama==0.4.1
+    # via pytest
+coverage==4.5.4
+    # via pytest-cov
 faker==2.0.1
-filelock==3.0.12          # via tox
-importlib-metadata==0.19  # via pluggy, pytest, tox
-more-itertools==7.2.0     # via pytest, zipp
-packaging==19.1           # via pytest, tox
-pluggy==0.12.0            # via pytest, tox
-py==1.8.0                 # via pytest, tox
+    # via -r requirements-test.in
+filelock==3.0.12
+    # via tox
+importlib-metadata==0.19
+    # via tox
+more-itertools==7.2.0
+    # via pytest
+packaging==19.1
+    # via
+    #   pytest
+    #   tox
+pluggy==0.13.1
+    # via
+    #   pytest
+    #   tox
+py==1.8.0
+    # via
+    #   pytest
+    #   tox
 pyfakefs==3.6
-pyparsing==2.4.2          # via packaging
-pytest-cov==2.7.1
+    # via -r requirements-test.in
+pyparsing==2.4.2
+    # via packaging
 pytest==5.1.2
-python-dateutil==2.8.0    # via faker
-six==1.12.0               # via faker, packaging, python-dateutil, tox
-text-unidecode==1.2       # via faker
-toml==0.10.0              # via tox
+    # via
+    #   -r requirements-test.in
+    #   pytest-cov
+pytest-cov==2.7.1
+    # via -r requirements-test.in
+python-dateutil==2.8.0
+    # via faker
+six==1.12.0
+    # via
+    #   faker
+    #   packaging
+    #   python-dateutil
+    #   tox
+text-unidecode==1.2
+    # via faker
+toml==0.10.0
+    # via tox
 tox==3.13.2
-virtualenv==16.7.4        # via tox
-wcwidth==0.1.7            # via pytest
-zipp==0.6.0               # via importlib-metadata
+    # via -r requirements-test.in
+virtualenv==16.7.4
+    # via tox
+wcwidth==0.1.7
+    # via pytest
+zipp==0.6.0
+    # via importlib-metadata
