+++ requirements.txt
@@ -1,17 +1,27 @@
 #
-# This file is autogenerated by pip-compile
+# This file is autogenerated by pip-compile with python 3.9
 # To update, run:
 #
 #    pip-compile --output-file=requirements.txt requirements.in
 #
-certifi==2019.6.16        # via requests
-chardet==3.0.4            # via requests
-fuzzywuzzy==0.17.0
-idna==2.8                 # via requests
+certifi==2021.5.30
+    # via requests
+charset-normalizer==2.0.4
+    # via requests
+fuzzywuzzy==0.18.0
+    # via -r requirements.in
+idna==3.2
+    # via requests
 omdbpy==0.1.14
-python-levenshtein==0.12.0
-requests==2.22.0
-urllib3==1.25.3           # via requests
+    # via -r requirements.in
+python-levenshtein==0.12.2
+    # via -r requirements.in
+requests==2.26.0
+    # via
+    #   -r requirements.in
+    #   omdbpy
+urllib3==1.26.6
+    # via requests
 
 # The following packages are considered to be unsafe in a requirements file:
-# setuptools==41.2.0        # via python-levenshtein
+# setuptools
