+++ tox.ini
@@ -4,7 +4,7 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py36, py37, lint, build
+envlist = py39, lint, build
 skip_missing_interpreters = true
 
 [testenv]
@@ -15,8 +15,7 @@ commands = {[testenv:unit]commands}
 
 [travis]
 python =
-    3.6: py36
-    3.7: py37, lint, build
+    3.9: py39, lint, build
 
 ; Trying this from https://github.com/Knewton/python-project-template-with-pip-compile/blob/master/tox.ini
 [testenv:pip-compile]
@@ -137,7 +136,7 @@ line_length=1000
 default_section = THIRDPARTY
 known_first_party = movie_file_fixer
 sections = FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
-exclude = .tox,env
+; #exclude = .tox,env
 
 [report]
 omit =
